test: [NUTR-118] - refactor tests
