--- src/app/modules/Auth/controller/index.ts
@@ -1,5 +1,0 @@
-import AuthController from './AuthController';
-
-export type optionsController = 'signUp' | 'signIn';
-
-export default AuthController;
